refactor(opencode-fork): lift profiles, delete legacy executor.ts
Phase 7 part 2. The src/engine/executor.ts shell has been a re-export
layer since Phase 1; with screen + classifier off the SDK there's no
reason to keep the indirection.

- src/engine/profiles.ts (new): ExecutorConfig, ExecutionResult,
  GitAccessProfile, GitSandboxAccess, GITHUB_PERMISSION_PROFILES,
  loadAgentContext — the only things callers actually needed from
  executor.ts.
- src/engine/opencode-executor.ts: imports types from ./profiles.js
  directly.
- src/workflows/{runner,simple,resume}.ts + src/engine/chat.ts:
  imports rerouted to ./profiles.js. Runner pulls executeAgent from
  opencode-executor.js directly.
- src/workflows/runner.test.ts: vi.mock target switched to
  ../engine/opencode-executor.js.
- src/engine/executor.ts: deleted.

npx tsc --noEmit clean; npx vitest run 387 passed / 1 todo.
