+++ src/workflows/resume.ts
@@ -1,5 +1,5 @@
 import type { StateDb, WorkflowRun } from "../state/db.js";
-import type { ExecutorConfig } from "../engine/executor.js";
+import type { ExecutorConfig } from "../engine/profiles.js";
 import type { GitHubClient } from "../engine/github.js";
 import type { ModelConfig } from "../config.js";
 import { runWorkflow, type ApprovalGateConfig, type RunnerCallbacks } from "./runner.js";
